Add a script to simplify updating a branch from the remote
This script is, effectively, a small wrapper around three git commands
that I have used for quite some time to update a branch based on the
remote branch. This script is designed to simplify the process and to
make it more mature, less prone to error, and without the need to
specifically remember.